requirements.txt
google-generativeai>=0.3.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pandas>=2.0.0
tqdm>=4.65.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
